Move EventPipeInternal into a separate file.
