+++ src/vm/mscorlib.cpp
@@ -84,10 +84,12 @@
 #if defined(FEATURE_EVENTSOURCE_XPLAT)
 #include "nativeeventsource.h"
 #include "eventpipe.h"
+#include "eventpipeinternal.h"
 #endif //defined(FEATURE_EVENTSOURCE_XPLAT)
 
 #ifdef FEATURE_PERFTRACING
 #include "eventpipe.h"
+#include "eventpipeinternal.h"
 #endif //FEATURE_PERFTRACING
 
 #endif // CROSSGEN_MSCORLIB
